requirements.txt

fastapi
uvicorn
firebase-admin
python-dotenv
jinja2
pydantic[email]
